Remove use of channels within the metrics collector struct
There's no need to use channels in the metrics collector struct; when
the metric reporting functions are called, we can update the metric
storage maps straight away. At the moment, we're just pushing to a
channel, and then receiving from it and immediately writing to the map -
the channels serve no purpose.

Also, remove the error label from the resource error metric, as we've
decided there's no need to differentiate between get and renewal errors.