Define resource process metrics for total, success and errors
These metrics track succeses, errors and totals for when resources are
"processed". Processing resources is the step that happens after a
resource is fetched from Vault - the resource is written to disk, and
then, if there's an exec specified on the resource, the given command is
executed.

The rationale behind these new metrics is to give visibility on when the
disk write or exec parts of resources fails.